app/requirements.txt
requests==2.24.0
Pillow==8.2.0
Flask==1.1.2
torch==1.8.0
torchvision==0.9.0
wget==3.2